+++ Werewolf for Telegram/Werewolf Node/Models/CustomGifData.cs
@@ -11,7 +11,7 @@ namespace Werewolf_Node.Models
         public bool HasPurchased { get; set; }
         public bool? Approved { get; set; } = null;
         public bool NSFW { get; set; }
-        public int ApprovedBy { get; set; }
+        public long ApprovedBy { get; set; }
         public string VillagerDieImage { get; set; }
         public string WolfWin { get; set; }
         public string WolvesWin { get; set; }
